uploading files/requirements.txt
flask==3.0.2
flask-cors==4.0.0
PyMuPDF==1.23.26
pytesseract==0.3.10
Pillow==10.2.0
Werkzeug==3.0.1 